Life time trigger and key trigger now is able to trigger on any key (#63)
* Added lifetime events

* fixed field value when trigger has no fields

* keytrigger can now trigger on any key

* cleaned up code

+++ src/event_trigger_runner/default_triggers/KeyTriggerFactory.cpp
@@ -11,7 +11,7 @@ std::unique_ptr<Trigger> KeyTriggerFactory::create(std::string name, const Dynam
 	
 	return std::make_unique<KeyTrigger>(
 		name, 
-		(unsigned short) trigger_key.get_value()
+		(int) trigger_key.get_value()
 	);
 }
